+++ CHANGELOG.md
@@ -8,11 +8,16 @@
   - It represents a value in `ticks` that the player must wait to reopen the Loot Ball.
   - A value of `0` disables this behaviour.
   - A small value (like `1`) indicates that the player can reopen the Loot Ball indefinitely.
-- Added new configs for some Loot Ball defaults:
+- Added new configs:
   - `loot_ball.defaults.uses`: Default amount of uses a Loot Ball has. Defaults to `1`.
   - `loot_ball.defaults.multiplier`: Default loot multiplier for Loot Balls. Defaults to `1.0`.
   - `loot_ball.defaults.player_timer`: Default player timer in ticks. Defaults to `0`.
   - `loot_ball.defaults.despawn_tick`: Default despawn tick timer. Defaults to `0`.
+  - `loot_ball.survival.drop.enabled`: Enable Loot Balls to drop as decorative items in survival mode. Defaults to `true`.
+- Added the ability for Loot Balls to drop as decorative items in survival mode.
+  - When a loot ball has no remaining uses, you can destroy them in survival and the loot ball will drop as a decorative item.
+  - This change is currently work in progress, so expect some bugs.
+  - You can disable this with the new `loot_ball.survival.drop.enabled` config.
 
 ## Fixes
 - Fixed wardens no longer being able to aggro loot balls